bump minimum version of RuboCop to 0.63.1
For the 0.10.0 release, the minimum version was 0.50 but due to
CI not checking specific versions, this was never really tested
and 0.63.1 was used as per the Travis CI logs [1].

For the 0.11.0 release, the minimum version was incorrectly set
to 0.49.1 as per #54 but again never really tested.

The CI is green with 0.63.1 so let's use that as minimum version.

[1] https://travis-ci.org/github/prontolabs/pronto-rubocop/builds/488138427